pyproject.toml
[project]
name = "siclo-scraping"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "dotenv>=0.9.9",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "openai>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pytest-playwright>=0.7.1",
]
